Make tasks table generic
